# api/openapi.yaml
openapi: 3.0.3
info:
  title: SCORM CMI API
  description: |
    A comprehensive REST API for managing SCORM 2004 courses and CMI (Computer Managed Instruction) data.
    
    This API provides endpoints for:
    - Uploading and managing SCORM packages
    - Creating and managing learner registrations
    - Tracking CMI data (progress, scores, interactions)
    - Generating learning analytics reports
    
    The API follows SCORM 2004 4th Edition specifications and supports all major CMI data elements.
  version: 1.0.0
  contact:
    name: SCORM CMI API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
paths:
...
  /registrations:
    get:
      summary: List registrations
      description: Retrieve a list of learner registrations, optionally filtered by learner
      tags:
        - Registrations
      parameters:
        - name: learner_id
          in: query
          required: false
          schema:
            type: string
          description: Filter registrations by learner ID
      responses:
        '200':
          description: List of registrations retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    registration:
                      $ref: '#/components/schemas/Registration'
                    course_title:
                      type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create a new registration
      description: Enroll a learner in a course
      tags:
        - Registrations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationCreateRequest'
      responses:
        '201':
          description: Registration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Registration already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /cmi/{registrationId}:
    get:
      summary: Get CMI data
      description: Retrieve CMI (Computer Managed Instruction) data for a registration
      tags:
        - CMI Data
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier of the registration
        - name: element
          in: query
          required: false
          schema:
            type: string
          description: Specific CMI element to retrieve (e.g., "cmi.core.lesson_status")
      responses:
        '200':
          description: CMI data retrieved successfully
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CMIData'
                  - $ref: '#/components/schemas/CMIElement'
        '404':
          description: Registration or CMI element not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Set CMI data
      description: Update CMI data for a registration
      tags:
        - CMI Data
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier of the registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CMIUpdateRequest'
      responses:
        '200':
          description: CMI data updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  element:
                    type: string
                  value:
                    type: string
                  status:
                    type: string
                    example: "success"
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Registration:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier
        course_id:
          type: string
          description: Associated course ID
        learner_id:
          type: string
          description: Learner's unique identifier
        learner_name:
          type: string
          description: Learner's display name
        status:
          type: string
          enum: ["not_attempted", "incomplete", "completed", "failed", "passed", "browsed"]
          description: Current lesson status
        score:
          type: number
          nullable: true
          description: Raw score
        min_score:
          type: number
          nullable: true
          description: Minimum passing score
        max_score:
          type: number
          nullable: true
          description: Maximum possible score
        session_time:
          type: string
          description: Time spent in current session
        total_time:
          type: string
          description: Total time spent on course
        location:
          type: string
          description: Bookmark/location in course
        suspend_data:
          type: string
          description: Suspended session data
        exit:
          type: string
          description: Exit condition
        entry:
          type: string
          description: Entry condition
        created_at:
          type: string
          format: date-time
          description: Registration creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: Completion timestamp
      required:
        - id
        - course_id
        - learner_id
        - learner_name
        - status
        - created_at
        - updated_at
    RegistrationCreateRequest:
      type: object
      properties:
        course_id:
          type: string
          description: Course to enroll in
        learner_id:
          type: string
          description: Learner's unique identifier
        learner_name:
          type: string
          description: Learner's display name
      required:
        - course_id
        - learner_id
        - learner_name
...
    CMIData:
      type: object
      additionalProperties:
        oneOf:
          - type: string
          - type: number
          - type: object
            properties:
              value:
                type: string
              timestamp:
                type: string
                format: date-time
      description: Collection of CMI data elements
    CMIElement:
      type: object
      properties:
        element:
          type: string
          description: CMI element name
        value:
          type: string
          description: CMI element value
        timestamp:
          type: string
          format: date-time
          description: When this value was set
      required:
        - element
        - value
    CMIUpdateRequest:
      type: object
      properties:
        element:
          type: string
          description: CMI element to update (e.g., "cmi.core.lesson_status")
        value:
          type: string
          description: New value for the element
      required:
        - element
        - value
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...